pyproject.toml
[project]
name = "mu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "black>=24.10.0",
    "fastapi[standard]>=0.115.5",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.10.0",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
    "uvicorn>=0.32.1",
    "strawberry-graphql[fastapi]>=0.251.0",
    "greenlet>=3.1.1",
]
